Refoctored routes, request now don't need to pass the user id

+++ src/main/kotlin/com/prmcoding/routes/comment/DeleteCommentRoute.kt
@@ -2,10 +2,9 @@ package com.prmcoding.routes.comment
 
 import com.prmcoding.data.requests.DeleteCommentRequest
 import com.prmcoding.responses.BasicApiResponse
-import com.prmcoding.routes.ifEmailBelongToUserId
+import com.prmcoding.routes.userId
 import com.prmcoding.service.CommentService
 import com.prmcoding.service.LikeService
-import com.prmcoding.service.UserService
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
@@ -15,7 +14,6 @@ import io.ktor.server.routing.*
 
 fun Route.deleteCommentRoute(
     commentService: CommentService,
-    userService: UserService,
     likeService: LikeService
 ) {
     authenticate {
@@ -26,27 +24,20 @@ fun Route.deleteCommentRoute(
                     return@delete
                 }
 
-            ifEmailBelongToUserId(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongsToUserId
-            ) {
-                val isDeletedComment = commentService.deleteComment(commentId = request.commentId)
-                if (isDeletedComment) {
-                    likeService.deleteLikesForParent(parentId = request.commentId)
-                    call.respond(
-                        status = HttpStatusCode.OK,
-                        message = BasicApiResponse(
-                            successful = true
-                        )
-                    )
-                } else {
-                    call.respond(
-                        status = HttpStatusCode.NotFound,
-                        message = BasicApiResponse(
-                            successful = false
-                        )
-                    )
-                }
+
+            val comment = commentService.getCommentById(commentId = request.commentId)
+
+            if (comment?.userId != call.userId) {
+                call.respond(HttpStatusCode.Unauthorized)
+                return@delete
+            }
+
+            val isDeletedComment = commentService.deleteComment(commentId = request.commentId)
+            if (isDeletedComment) {
+                likeService.deleteLikesForParent(parentId = request.commentId)
+                call.respond(HttpStatusCode.OK, BasicApiResponse(successful = true))
+            } else {
+                call.respond(HttpStatusCode.NotFound, BasicApiResponse(successful = false))
             }
         }
     }
